requirements.txt
altair=5.0.1
anyio=4.6.2
argon2-cffi-bindings=21.2.0
argon2-cffi=21.3.
arrow-cpp=16.1.0
asttokens=2.0.5
async-lru=2.0.4
attrs=24.2.0
aws-c-auth=0.6.19
aws-c-cal=0.5.20
aws-c-common=0.8.5
aws-c-compression=0.2.16
aws-c-event-stream=0.2.15
aws-c-http=0.6.25
aws-c-io=0.13.10
aws-c-mqtt=0.7.13
aws-c-s3=0.1.51
aws-c-sdkutils=0.1.6
aws-checksums=0.1.13
aws-crt-cpp=0.18.16
aws-sdk-cpp=1.10.55
babel=2.11.0
beautifulsoup4=4.12.3
blas=1.0
bleach=6.2.0
blinker=1.6.2
boost-cpp=1.82.0
bottleneck=1.4.2
brotli-bin=1.0.9
brotli-python=1.0.9
brotli=1.0.9
bzip2=1.0.8
c-ares=1.19.1
ca-certificates=2024.12.31
cachetools=5.3.3
certifi=2024.12.14
cffi=1.17.1
charset-normalizer=3.3.2
click=8.1.7
colorama=0.4.6
comm=0.2.1
contourpy=1.3.1
cycler=0.11.
debugpy=1.6.7
decorator=5.1.1
defusedxml=0.7.1
et_xmlfile=1.1.0
executing=0.8.3
expat=2.6.3
fonttools=4.51.0
freetype=2.12.1
gflags=2.2.2
gitdb=4.0.7
gitpython=3.1.43
glog=0.5.0
h11=0.14.0
httpcore=1.0.2
httpx=0.27.0
icu=73.1
idna=3.7
intel-openmp=2023.1.0
ipykernel=6.29.5
ipython=8.27.0
jedi=0.19.1
jinja2=3.1.4
jpeg=9e=h827c3e9_3
json5=0.9.25
jsonschema-specifications=2023.7.1
jsonschema=4.23.0
jupyter-lsp=2.2.0
jupyter_client=8.6.0
jupyter_core=5.7.2
jupyter_events=0.10.0
jupyter_server=2.14.1
jupyter_server_terminals=0.4.4
jupyterlab=4.2.5
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.27.3
kiwisolver=1.4.4
krb5=1.20.1=h5b6d351_0
lasio=0.31
lcms2=2.12=h83e58a3_0
lerc=3.0
libabseil=20240116.2
libboost=1.82.0
libbrotlicommon=1.0.9
libbrotlidec=1.0.9
libbrotlienc=1.0.9
libclang13=14.0.6
libclang=14.0.6
libcurl=8.9.1
libdeflate=1.17
libevent=2.1.12
libffi=3.4.4
libgrpc=1.62.2
libpng=1.6.39
libpq=17.0
libprotobuf=4.25.3
libsodium=1.0.18
libssh2=1.11.1
libthrift=0.15.0
libtiff=4.5.1
libwebp-base=1.3.2
lz4-c=1.9.4
markdown-it-py=2.2.0
markupsafe=2.1.3
matplotlib-base=3.9.2
matplotlib-inline=0.1.6
matplotlib=3.9.2
mdurl=0.1.0
mistune=2.0.4
mkl-service=2.4.0
mkl=2023.1.0
mkl_fft=1.3.11
mkl_random=1.2.8
nbclient=0.8.0
nbconvert=7.16.4
nbformat=5.10.4
nest-asyncio=1.6.0
notebook-shim=0.2.3
notebook=7.2.2
numexpr=2.10.1
numpy-base=1.26.4
numpy=1.26.4
openjpeg=2.5.2
openpyxl=3.1.5
openssl=3.0.15
orc=2.0.1
overrides=7.4.0
packaging=24.1
pandas=2.2.3
pandocfilters=1.5.
parso=0.8.3
pillow=11.0.0
pip=24.2
platformdirs=3.10.0
plotly=5.24.1
ply=3.11
prometheus_client=0.21.0
prompt-toolkit=3.0.43
prompt_toolkit=3.0.43
protobuf=4.25.3
psutil=5.9.0
pure_eval=0.2.2
pyarrow=16.1.0
pycparser=2.21
pydeck=0.8.0
pygments=2.15.1
pyparsing=3.2.0
pyqt5-sip=12.13.0
pyqt=5.15.10
pysocks=1.7.1
python-dateutil=2.9.0post0
python-fastjsonschema=2.20.0
python-json-logger=2.0.7
python-tzdata=2023.3
python=3.12.7
pytz=2024.1
pywin32=305
pywinpty=2.0.10
pyyaml=6.0.2
pyzmq=25.1.2
qt-main=5.15.2
re2=2022.04.01
referencing=0.30.2
requests=2.32.3
rfc3339-validator=0.1.4
rfc3986-validator=0.1.1
rich=13.9.4
rpds-py=0.10.6
scipy=1.15.1
seaborn=0.13.2
send2trash=1.8.2
setuptools=72.1.0
sip=6.7.12
six=1.16.0
smmap=4.0.
snappy=1.2.1
sniffio=1.3.0
soupsieve=2.5
sqlite=3.45.3
stack_data=0.2.
streamlit=1.40.1
striplog=0.9.2
tbb=2021.8.0
tenacity=9.0.0
terminado=0.17.1
tinycss2=1.2.1
tk=8.6.14
toml=0.10.2
toolz=0.12.0
tornado=6.4.1
tqdm=4.67.1
traitlets=5.14.3
typing-extensions=4.11.0
typing_extensions=4.11.0
unicodedata2=15.1.0
urllib3=2.2.3
utf8proc=2.6.1
vc=14.40
vs2015_runtime=14.40.33807
watchdog=4.0.1
wcwidth=0.2.5
webencodings=0.5.1
websocket-client=1.8.0
wellpathpy=0.5.1
welly=0.5.2
wheel=0.44.0
win_inet_pton=1.1.0
winpty=0.4.3=4
xz=5.4.6
yaml=0.2.5
zeromq=4.3.5
zlib=1.2.13
zstd=1.5.6
